+++ hub/api/dataset.py
@@ -14,7 +14,7 @@ from hub.core.chunk_engine.write import write_array, write_dataset_meta
 from typing import Union, Dict, Optional
 import numpy as np
 import warnings
-from hub.api.pt import _to_pytorch
+from hub.api.pytorch import _to_pytorch
 
 
 class Dataset:
